Updated Analytics Provider to Segment
