+++ package.json
@@ -20,6 +20,7 @@
     "@babel/plugin-transform-react-jsx": "^7.13.12",
     "@fingerprintjs/fingerprintjs-pro": "^3.5.3",
     "@n8tb1t/use-scroll-position": "^2.0.3",
+    "@segment/analytics-next": "^1.46.0",
     "@storybook/addon-controls": "^6.3.7",
     "@storybook/cli": "^6.3.7",
     "@styled-system/theme-get": "^5.1.2",
