+++ providers/AppProvider.js
@@ -3,7 +3,12 @@ import PropTypes from 'prop-types';
 import { ApolloProvider } from '@apollo/client';
 
 import { useApollo } from 'lib/apolloClient';
-import { AnalyticsProvider, AuthProvider, ModalProvider, GroupFiltersProvider } from 'providers';
+import {
+  AnalyticsProvider,
+  AuthProvider,
+  ModalProvider,
+  GroupFiltersProvider,
+} from 'providers';
 import { ModalManager } from 'providers/ModalProvider';
 import modals from 'config/modals';
 import { ThemeProvider } from 'ui-kit';
@@ -13,7 +18,7 @@ function AppProvider(props = {}) {
   return (
     <ApolloProvider client={apolloClient}>
       <AuthProvider>
-        <AnalyticsProvider>
+        <AnalyticsProvider writeKey={process.env.SEGMENT_KEY}>
           <ThemeProvider>
             <GroupFiltersProvider>
               <ModalProvider modals={modals}>
